+++ airo-robots/airo_robots/manipulators/__init__.py
@@ -1,2 +1,3 @@
+from airo_robots.manipulators.hardware.realman import RealmanControl  # noqa F401
 from airo_robots.manipulators.hardware.ur_rtde import URrtde  # noqa F401
 from airo_robots.manipulators.position_manipulator import PositionManipulator  # noqa F401